merged celery and celery_powder together
Query: in the Edible Vegetable class, should the instances 'Celery Powder' and 'Celery' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Celery powder and celery are semantic and ontologic duplicates as both refer to the same vegetable.
meaning that instances celery and celery_powder are semantically identical.
Query cache in files:
- /kgfiller/data/cache-57a18742e1e1e6763c47a6b46573d5c8c6d07ae7385ec35deefb99566cf07ff3.yml