merged peppers and bell_peppers together
Query: in the Edible Vegetable class, should the instances 'Bell peppers' and 'Peppers' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Bell peppers and Peppers are semantic and ontologic duplicates. They refer to the same vegetable.
meaning that instances peppers and bell_peppers are semantically identical.
Query cache in files:
- /kgfiller/data/cache-a0967670aaf63ab08492afd7c682a475c929bfa23d5d07d664819dbbdcbcb660.yml