merged peppers and green_pepper together
Query: in the Edible Vegetable class, should the instances 'Green Pepper' and 'Peppers' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Green Pepper and Peppers are semantic and ontologic duplicates. They are the same vegetable with different names.
meaning that instances peppers and green_pepper are semantically identical.
Query cache in files:
- /kgfiller/data/cache-3bd619900beacf99d167fac105c851bf3b5f7fb074af6fcfa63d56376350d29e.yml