merged tomato and tomato_paste together
Query: in the Edible Vegetable class, should the instances 'Tomato paste' and 'Tomato' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Tomato paste and tomato are semantic and ontologic duplicates. They are both referring to the same fruit, tomato.
meaning that instances tomato and tomato_paste are semantically identical.
Query cache in files:
- /kgfiller/data/cache-fbf7fd83bcd1b638f36ed45ae4e59512ac509356f5db68a14fd05aa33d35b15f.yml