instances vegetables and vegetable_oil were NOT merged together
Query: in the Edible Vegetable class, should the instances 'Vegetable oil' and 'Vegetables' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

Vegetable oil and vegetables are not semantic and ontologic duplicates. They are different concepts. Vegetable oil is a fat extracted from vegetables, while vegetables are a group of plant products that are eaten as part of a healthy diet.
meaning that instances vegetables and vegetable_oil are different.
Query cache in files:
- /kgfiller/data/cache-d630f6a9251e07ca20ae270b7b74f7ed369ec1185e65790d39b59bd380deee76.yml