merged onion and diced_onion together
Query: in the Edible Vegetable class, should the instances 'Diced onion' and 'Onion' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Diced onion and onion are semantic and ontologic duplicates. They are both referring to the same vegetable.
meaning that instances onion and diced_onion are semantically identical.
Query cache in files:
- /kgfiller/data/cache-b28354afefdc7c2f13ebee2f66379216e7483dbb5ae29820d3a608675928bd79.yml