merged vegetable_broth and bouillon_broth together
Query: in the Edible Vegetable class, should the instances 'Bouillon broth' and 'Vegetable broth' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Bouillon broth and vegetable broth are semantic and ontologic duplicates, as they refer to the same concept.
meaning that instances vegetable_broth and bouillon_broth are semantically identical.
Query cache in files:
- /kgfiller/data/cache-6f808e4344796fd541d577d048cef6728c03718e7020ee61fca360a142f271e6.yml