instances peppercorns and cornstarch were NOT merged together
Query: in the Edible Vegetable class, should the instances 'Cornstarch' and 'Peppercorns' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

"Cornstarch" and "Peppercorns" are not semantic and ontologic duplicates. They are two different food ingredients with distinct nutritional profiles and culinary uses.
meaning that instances peppercorns and cornstarch are different.
Query cache in files:
- /kgfiller/data/cache-166d53b98b28029d300bea60297c27b0fb961c81139fb288a48fb0cec31dd447.yml