instances peppercorns and corn were NOT merged together
Query: in the Edible Vegetable class, should the instances 'Corn' and 'Peppercorns' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Corn and peppercorns are both semantic and ontologic duplicates. They are both vegetables and are often referred to as the same thing.
meaning that instances peppercorns and corn are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-2872de264aacdf9ab366bf3eb3490b1d07a5065826a88f28bbeed600649046b2.yml