instances beans and bean_sprouts were NOT merged together
Query: in the Edible Vegetable class, should the instances 'Bean sprouts' and 'Beans' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No. Bean sprouts and beans are not semantic and ontologic duplicates.

Bean sprouts are germinated beans that are eaten as a different vegetable from beans.
meaning that instances beans and bean_sprouts are different.
Query cache in files:
- /kgfiller/data/cache-5102831f6e36f67390eae0b0e9ba7714266b815e99cacce9b3299e04a05df81a.yml