merged wine and shaoxing_wine together
Query: in the Alcoholic Drink class, should the instances 'Shaoxing wine' and 'Wine' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	yes

Shaoxing wine and Wine are semantic and ontologic duplicates as they both refer to the same beverage.
meaning that instances wine and shaoxing_wine are semantically identical.
Query cache in files:
- /kgfiller/data/cache-a8dea66ceb9d4d98955a0141234c71d3b91f4c2adf86ff1dce4249a983a395fb.yml