instances green_tea and black_tea were NOT merged together
Query: in the Tea class, should the instances 'Black tea' and 'Green tea' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

Black tea and green tea are different types of tea with different chemical compositions and health effects. They should not be merged together as semantic and ontologic duplicates.
meaning that instances green_tea and black_tea are different.
Query cache in files:
- /kgfiller/data/cache-4be51da6e7de8cddc032ba96003b232f3510c26b35d836c9e67481a42555625a.yml