instances green_tea and matcha_tea were NOT merged together
Query: in the Tea class, should the instances 'Green tea' and 'Matcha tea' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

Green tea and matcha tea are not semantic and ontologic duplicates. They are two different types of tea with distinct properties and benefits.
meaning that instances green_tea and matcha_tea are different.
Query cache in files:
- /kgfiller/data/cache-afe47ed644bfc144b441a50321881a9e9c903ad2dc18392fe5c792aca9f4e115.yml