instances black_tea and pu_erh_tea were NOT merged together
Query: in the Tea class, should the instances 'Black tea' and 'Pu-erh tea' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

Black tea and Pu-erh tea are not semantic and ontologic duplicates. They are two different types of tea with distinct characteristics and health benefits.
meaning that instances black_tea and pu_erh_tea are different.
Query cache in files:
- /kgfiller/data/cache-e60ecc4e9b27b0d560d8b22852ea5a767c346943e2549c8aefe30a4f7f8d0cc4.yml