instances rooibos_tea and matcha_tea were NOT merged together
Query: in the Tea class, should the instances 'Matcha tea' and 'Rooibos tea' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

Matcha tea and Rooibos tea are not semantic and ontologic duplicates. They are different types of tea with distinct nutritional profiles.
meaning that instances rooibos_tea and matcha_tea are different.
Query cache in files:
- /kgfiller/data/cache-f35d552485795e8c9386075a9b92d04c23264e7090bcf1918ddb096703ec2d4c.yml