instances jasmine_tea and flora_tea_petals were NOT merged together
Query: in the Tea class, should the instances 'Flora tea petals' and 'Jasmine tea' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

The instances 'Flora tea petals' and 'Jasmine tea' are not semantic and ontologic duplicates. They are different concepts. 'Flora tea petals' refers to the petals of flowers that are used in tea blends, while 'Jasmine tea' refers to a specific type of tea made from the leaves of the Jasminum sambacaucus plant.
meaning that instances jasmine_tea and flora_tea_petals are different.
Query cache in files:
- /kgfiller/data/cache-87d7cd83155949216e0a552956e5431cca393059caf2f935e7b533dca5f7cfdb.yml