merged juice and lime_juice together
Query: in the Juice class, should the instances 'Juice' and 'Lime juice' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Juice and lime juice are both instances of the class Juice. They are semantic and ontologic duplicates.
meaning that instances juice and lime_juice are semantically identical.
Query cache in files:
- /kgfiller/data/cache-208ad570d5be14bdb037a13bffaa433ae85f1e215ef81fc73ae758b98b42183d.yml