instances fruit_infused_water and rose_water were NOT merged together
Query: in the Drinking Water class, should the instances 'Fruit-Infused Water' and 'Rose water' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

Fruit-Infused Water and Rose Water are two distinct concepts. Fruit-Infused Water is a type of water that is flavored by fruits, while Rose Water is a type of flavored water that is made with rose petals.
meaning that instances fruit_infused_water and rose_water are different.
Query cache in files:
- /kgfiller/data/cache-cd233b747e23da24a44190f4d91f66471398f1c46288ad357acfa5294a5ff90a.yml