instances dried_thyme and dried_rosemary were NOT merged together
Query: in the Edible class, should the instances 'Dried rosemary' and 'Dried thyme' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Dried rosemary and dried thyme are both herbs that are in the same family, Lamiaceae, so they are semantic and ontologic duplicates.
meaning that instances dried_thyme and dried_rosemary are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-21a6cbf27966e93d79c117943cd34b418771dad08fb50d351c97fae5341aa2ef.yml