instances lactose and maltose were NOT merged together
Query: in the Sweetener class, should the instances 'Lactose' and 'Maltose' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	As a dietician, my role is to provide information and advice related to food and nutrition. In terms of the Sweetener class and the instances 'Lactose' and 'Maltose', they are not semantic or ontological duplicates, as they are two different types of sugars.

Lactose is a disaccharide sugar found in milk and dairy products. It is composed of glucose and galactose. Maltose, on the other hand, is also a disaccharide sugar but is found in grains, particularly in germinating grains. It is composed of two glucose molecules.

While both Lactose and Maltose are types of sugars and can be classified as sweeteners, they have different chemical structures, sources, and uses. Therefore, they should not be merged together as semantic or ontological duplicates.
meaning that instances lactose and maltose are different.
Query cache in files:
- /kgfiller/data/cache-995f87d9bdd0c26bf68552b62e020bba95de23b3bff98e3eb4778a12de6386fc.yml